Move verbose to util
