+++ go/util/verbose/verbose.go
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, version 2.0:
 // http://www.apache.org/licenses/LICENSE-2.0
 
-package config
+package verbose
 
 import (
 	flag "github.com/juju/gnuflag"
