+++ samples/go/nomdex/nomdex_update.go
@@ -16,6 +16,7 @@ import (
 	"github.com/attic-labs/noms/go/types"
 	"github.com/attic-labs/noms/go/util/profile"
 	"github.com/attic-labs/noms/go/util/status"
+	"github.com/attic-labs/noms/go/util/verbose"
 	"github.com/attic-labs/noms/go/walk"
 	humanize "github.com/dustin/go-humanize"
 	flag "github.com/juju/gnuflag"
@@ -41,7 +42,7 @@ func setupUpdateFlags() *flag.FlagSet {
 	flagSet.StringVar(&inPathArg, "in-path", "", "a value to search for items to index within ")
 	flagSet.StringVar(&outDsArg, "out-ds", "", "name of dataset to save the results to")
 	flagSet.StringVar(&relPathArg, "by", "", "a path relative to all the items in <in-path> to index by")
-	config.RegisterVerboseFlags(flagSet)
+	verbose.RegisterVerboseFlags(flagSet)
 	profile.RegisterProfileFlags(flagSet)
 	return flagSet
 }
